+++ common/util.py
@@ -1,0 +1,3 @@
+def info(msg):
+    """Print green information using ANSI code"""
+    print("\033[92m {}\033[00m".format(msg))
