--- matd3/common/util.py
@@ -1,17 +1,0 @@
-import tensorflow as tf
-
-
-def info(msg):
-    """Print green information using ANSI code"""
-    print("\033[92m {}\033[00m".format(msg))
-
-
-def clip_by_local_norm(gradients, norm):
-    """
-    Clips gradients by their own norm, NOT by the global norm
-    as it should be done (according to TF documentation).
-    This here is the way MADDPG does it.
-    """
-    for idx, grad in enumerate(gradients):
-        gradients[idx] = tf.clip_by_norm(grad, norm)
-    return gradients
