+++ matd3/util.py
@@ -1,0 +1,12 @@
+import tensorflow as tf
+
+
+def clip_by_local_norm(gradients, norm):
+    """
+    Clips gradients by their own norm, NOT by the global norm
+    as it should be done (according to TF documentation).
+    This here is the way MADDPG does it.
+    """
+    for idx, grad in enumerate(gradients):
+        gradients[idx] = tf.clip_by_norm(grad, norm)
+    return gradients
